requirements.txt
python-rtmidi==1.4.6
pyvjoy==1.0.1
